+++ ToDoApp.Buisiness/Services/TagService.cs
@@ -5,6 +5,7 @@ using TodoApp.Data.Interfaces;
 using TodoApp.Buisiness.Models;
 using AutoMapper;
 using TodoApp.Data.Models;
+using System;
 
 namespace TodoApp.Buisiness.Services
 {
@@ -21,18 +22,37 @@ namespace TodoApp.Buisiness.Services
 
         public async Task<int> Create(TagVO tag)
         {
+            if (await _dataProvider.IsDuplicate(_mapper.Map<TagDAO>(tag)))
+            {
+                throw new ArgumentException("A tag with the name " + tag.Name + " already exists");
+            }
             int createdId  = await _dataProvider.Create(_mapper.Map<TagDAO>(tag));
             return createdId;
         }
 
         public async Task Delete(int id)
         {
-            await _dataProvider.Delete(id);
+            if (await _dataProvider.Exists(id))
+            {
+                await _dataProvider.Delete(id);
+            }
+            else
+            {
+                throw new KeyNotFoundException();
+            }
+            
         }
 
         public async Task<TagVO> Get(int? id)
         {
-            return _mapper.Map<TagVO>(await _dataProvider.Get(id));
+            if (await _dataProvider.Exists(id))
+            {
+                return _mapper.Map<TagVO>(await _dataProvider.Get(id));
+            }
+            else
+            {
+                throw new KeyNotFoundException();
+            }
         }
 
         public async Task<IEnumerable<TagVO>> GetAll()
@@ -42,7 +62,14 @@ namespace TodoApp.Buisiness.Services
 
         public async Task Update(TagVO tag)
         {
-            await _dataProvider.Update(_mapper.Map<TagDAO>(tag));
+            if (await _dataProvider.Exists(tag.Id))
+            {
+                await _dataProvider.Update(_mapper.Map<TagDAO>(tag));
+            }
+            else
+            {
+                throw new KeyNotFoundException();
+            }
         }
     }
 }
